+++ src/middleware.ts
@@ -15,9 +15,9 @@ export async function middleware(request: NextRequest) {
   const accessToken = request.cookies.get('accessToken')?.value;
   const refreshToken = request.cookies.get('refreshToken')?.value;
 
-  if (!isPublicPath && !accessToken && !refreshToken) {
-    return NextResponse.redirect(new URL('/signin', request.url));
-  }
+  // if (!isPublicPath && !accessToken && !refreshToken) {
+  //   return NextResponse.redirect(new URL('/signin', request.url));
+  // }
 
   if (isPublicPath && accessToken && refreshToken) {
     return NextResponse.redirect(new URL('/stage', request.url));
